# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Parvath FinServ CRM",
    "version": "1.0.0",
    "description": "All business endpoints require a workspace-scoped session. Mutations require the CSRF token returned by /auth/csrf or /auth/login. Dates use Asia/Kolkata; money is an integer string in paise. See docs/api.md for lifecycle contracts."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "parvath.sid"
      }
    },
...
  },
  "paths": {
    "/clients": {
      "get": {
        "summary": "List authorized clients",
        "tags": [
          "clients"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clientId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success; JSON {data, meta?}"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Role or CSRF denied"
          },
          "404": {
            "description": "Record absent or outside workspace"
          },
          "409": {
            "description": "Duplicate, stale version or invalid transition"
          },
          "422": {
            "description": "Zod validation error with fieldErrors"
          },
          "503": {
            "description": "External integration not configured"
          }
        }
      },
      "post": {
        "summary": "Create clients",
        "tags": [
          "clients"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Client"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success; JSON {data, meta?}"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Role or CSRF denied"
          },
          "404": {
            "description": "Record absent or outside workspace"
          },
          "409": {
            "description": "Duplicate, stale version or invalid transition"
          },
          "422": {
            "description": "Zod validation error with fieldErrors"
          },
          "503": {
            "description": "External integration not configured"
          }
        }
      }
    },
...
    "/leads": {
      "get": {
        "summary": "List authorized leads",
        "tags": [
          "leads"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clientId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success; JSON {data, meta?}"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Role or CSRF denied"
          },
          "404": {
            "description": "Record absent or outside workspace"
          },
          "409": {
            "description": "Duplicate, stale version or invalid transition"
          },
          "422": {
            "description": "Zod validation error with fieldErrors"
          },
          "503": {
            "description": "External integration not configured"
          }
        }
      },
      "post": {
        "summary": "Create leads",
        "tags": [
          "leads"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Opportunity"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success; JSON {data, meta?}"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Role or CSRF denied"
          },
          "404": {
            "description": "Record absent or outside workspace"
          },
          "409": {
            "description": "Duplicate, stale version or invalid transition"
          },
          "422": {
            "description": "Zod validation error with fieldErrors"
          },
          "503": {
            "description": "External integration not configured"
          }
        }
      }
    },
...
    "/products": {
      "get": {
        "summary": "List authorized products",
        "tags": [
          "products"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clientId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success; JSON {data, meta?}"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Role or CSRF denied"
          },
          "404": {
            "description": "Record absent or outside workspace"
          },
          "409": {
            "description": "Duplicate, stale version or invalid transition"
          },
          "422": {
            "description": "Zod validation error with fieldErrors"
          },
          "503": {
            "description": "External integration not configured"
          }
        }
      },
      "post": {
        "summary": "Create products",
        "tags": [
          "products"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success; JSON {data, meta?}"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Role or CSRF denied"
          },
          "404": {
            "description": "Record absent or outside workspace"
          },
          "409": {
            "description": "Duplicate, stale version or invalid transition"
          },
          "422": {
            "description": "Zod validation error with fieldErrors"
          },
          "503": {
            "description": "External integration not configured"
          }
        }
      }
    },
...
    "/dashboard": {
      "get": {
        "summary": "Authoritative counts, attention, calendar and chart data",
        "tags": [
          "dashboard"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clientId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success; JSON {data, meta?}"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Role or CSRF denied"
          },
          "404": {
            "description": "Record absent or outside workspace"
          },
          "409": {
            "description": "Duplicate, stale version or invalid transition"
          },
          "422": {
            "description": "Zod validation error with fieldErrors"
          },
          "503": {
            "description": "External integration not configured"
          }
        }
      }
    },
    "/search": {
      "get": {
        "summary": "Authorized global search; q length 2+",
        "tags": [
          "search"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clientId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success; JSON {data, meta?}"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Role or CSRF denied"
          },
          "404": {
            "description": "Record absent or outside workspace"
          },
          "409": {
            "description": "Duplicate, stale version or invalid transition"
          },
          "422": {
            "description": "Zod validation error with fieldErrors"
          },
          "503": {
            "description": "External integration not configured"
          }
        }
      }
    },
    "/notifications": {
      "get": {
        "summary": "Current member notifications",
        "tags": [
          "notifications"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clientId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success; JSON {data, meta?}"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Role or CSRF denied"
          },
          "404": {
            "description": "Record absent or outside workspace"
          },
          "409": {
            "description": "Duplicate, stale version or invalid transition"
          },
          "422": {
            "description": "Zod validation error with fieldErrors"
          },
          "503": {
            "description": "External integration not configured"
          }
        }
      }
    },
...
  }
}